1. If present, body is now attached, functions return total length of http message 2. Body is now attached correctly

+++ httpMessageConvert.h
@@ -7,7 +7,7 @@
 
 int strToHTTP(HTTPREQUEST *buffer, char *strBuf);
 
-int HTTPTOstr(HTTPRESPONSE *buffer, char *strBuf);
+long int HTTPTOstr(HTTPRESPONSE *buffer, char *strBuf);
 
 
 #endif
